Bugfix: crash when editing frame after changing parameters
If you edit animation parameters and then try to edit the current frame,
the program will crash because the parameter box made a deep copy of the
current animation for temporary editing. When copying back after editing,
the new frames would take the place of the old ones. The frames are
identical, but the undo stack holds on to the old references which no
longer exist in the animation. By doing a shallow copy instead (which is
fine because the parameter editor doesn't edit frames) we can keep the old
referenced. I presume this also fixes a memory leak.